Modifications...
This isn't going to work oh well. In an attempt to understand the format of the input and ensure the parser works, I've tried modifying the token names, and so the lexical analyser. Also modified calcGoto() a bit to ensure it works correctly if no entry is already present in the map. Created calcSLRgoto().