Fix follow for $ & fix start prod for augmGram (IT WORKS!)
Moved adding the Inputendmarker to the grammar's start symbol's follow set to the method calcAllFollowSets() (it previously replaced the existing set by an empty one when calcAllFollowSets() was called). Used "INPUTENDMARKER" instead of "$" here and in the accept action. Fix new "S' -> S" production in creation of the augmented grammar. Drum roll... it now works!!!!!! :)